Restructure and redesign of uni list lab
