+++ app/components/MainTabBar.js
